Replace one impl with a derive
